fixes to radar preprocessing

+++ src/convert_radar_data_to_netCDF.py
@@ -9,17 +9,26 @@ import concurrent.futures
 
 
 def convert_all_grb_files_to_netCDF(grib_file_path, grib2_res_folder_path, nc_res_folder_path):
+    print(f"parsing file: {grib_file_path}")
     if not (grib2_res_folder_path / grib_file_path.parent.name).exists():
         os.mkdir(grib2_res_folder_path / grib_file_path.parent.name)
     grib_file_path_for_reading_using_xarray = grib2_res_folder_path / grib_file_path.parent.name / (
             grib_file_path.stem + ".grb2")
+    file_path_net_CDF = nc_res_folder_path / grib_file_path.parent.name / (grib_file_path.stem + ".nc")
+    if file_path_net_CDF.exists():
+        print(f"the netCDF file: {file_path_net_CDF} exists")
+        return
     if not grib_file_path_for_reading_using_xarray.exists():
         if grib_file_path.suffix != ".grb2":
             if grib_file_path.suffix == ".gz":
                 with gzip.open(grib_file_path, 'rb') as f_in:
                     uncompressed_data = f_in.read()
             elif grib_file_path.suffix == ".Z":
-                uncompressed_data = unlzw3.unlzw(grib_file_path)
+                try:
+                    uncompressed_data = unlzw3.unlzw(grib_file_path)
+                except ValueError as e:
+                    print(e)
+                    return
             else:
                 raise Exception(f"unknown extension: {grib_file_path}")
             with open(grib_file_path_for_reading_using_xarray, "wb") as f_out:
@@ -29,10 +38,8 @@ def convert_all_grb_files_to_netCDF(grib_file_path, grib2_res_folder_path, nc_re
     data = xarray.open_dataset(grib_file_path_for_reading_using_xarray, engine='cfgrib')
     if not (nc_res_folder_path / grib_file_path.parent.name).exists():
         os.mkdir(nc_res_folder_path / grib_file_path.parent.name)
-    file_path_net_CDF = nc_res_folder_path / grib_file_path.parent.name / (grib_file_path.stem + ".nc")
-    if not file_path_net_CDF.exists():
-        data.to_netcdf(file_path_net_CDF)
-        nc_file = netCDF4.Dataset(file_path_net_CDF)
+    data.to_netcdf(file_path_net_CDF)
+    nc_file = netCDF4.Dataset(file_path_net_CDF)
     print(f"finished with file: {grib_file_path}")
